requirements.txt
atlassian-python-api==3.28.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyGithub==1.56
PyJWT==2.5.0
PyNaCl==1.5.0
python-gitlab==2.10.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
six==1.16.0
tomli==1.2.3
urllib3==1.26.12
wrapt==1.14.1
